requirements.txt
asgiref==3.7.2
automium==0.2.6
automium-web==0.1.1
blinker==1.7.0
click==8.1.7
Django==4.2.7
djangorestframework==3.14.0
Flask==3.0.0
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
legacy==0.1.6
MarkupSafe==2.1.3
psycopg2==2.9.9
pytz==2023.3.post1
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
Werkzeug==3.0.1
